+++ packages/tsurlfilter/CHANGELOG.md
@@ -8,12 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: manually add compare links for version to the end of the file -->
 <!-- e.g. [1.0.77]: https://github.com/AdguardTeam/tsurlfilter/compare/tsurlfilter-v1.0.76...tsurlfilter-v1.0.77 -->
 
-## Unreleased
+
+## [2.0.5] - 2023-04-24
 
 ### Changed
 
 - RuleConverter speed has been improved by 2x [#83](https://github.com/AdguardTeam/tsurlfilter/issues/83)
 - Declarative converter build has been separated from the main package
+
+
+<!-- 2.0.4 is the same as 2.0.3, no difference -->
 
 
 ## [2.0.3] - 2023-04-17